calib sync
