--- pyhealth/calib/predictionset/predictionset/cluster/__init__.py
@@ -1,6 +1,0 @@
-"""Cluster-based and neighborhood prediction set methods"""
-
-from pyhealth.calib.predictionset.cluster.cluster_label import ClusterLabel
-from pyhealth.calib.predictionset.cluster.neighborhood_label import NeighborhoodLabel
-
-__all__ = ["ClusterLabel", "NeighborhoodLabel"]
