--- pyhealth/calib/predictionset/predictionset/covariate/__init__.py
@@ -1,8 +1,0 @@
-"""Covariate shift adaptive prediction set methods"""
-
-from pyhealth.calib.predictionset.covariate.covariate_label import (
-    CovariateLabel,
-    fit_kde,
-)
-
-__all__ = ["CovariateLabel", "fit_kde"]
